pyproject.toml
[tool.poetry]
name = "avatar-validator"
version = "0.0.1"
description = "validate an avatar"
authors = ["helmi nour"]
license = "GPL"
readme = "README.md"
packages = [{ include = "api" }]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
